end of lesson 22

+++ day22-node-and-express-pt-5-authentication-and-access-control/job-application-tracker-api/server.js
@@ -1,7 +1,9 @@
 const express = require("express");
 const app = express();
 const port = 4000;
-const { JobApplication } = require("./models");
+const bcrypt = require("bcryptjs");
+const session = require("express-session");
+const { JobApplication, User } = require("./models");
 require("dotenv").config();
 
 app.use((req, res, next) => {
@@ -13,13 +15,108 @@ app.use((req, res, next) => {
   next();
 });
 app.use(express.json());
+app.use(
+  session({
+    secret: process.env.SESSION_SECRET,
+    resave: false,
+    saveUninitialized: false,
+    cookie: {
+      maxAge: 3600000, // 1 hour
+    },
+  })
+);
+const authenticateUser = (req, res, next) => {
+  if (!req.session.userId) {
+    return res
+      .status(401)
+      .json({ message: "You must be logged in to view this page." });
+  }
+  next();
+};
 
 app.get("/", (req, res) => {
   res.send("Welcome to the Job App Tracker API!!!!");
 });
 
+app.post("/signup", async (req, res) => {
+  const hashedPassword = await bcrypt.hash(req.body.password, 10);
+
+  try {
+    const user = await User.create({
+      name: req.body.name,
+      email: req.body.email,
+      password: hashedPassword
+    })
+    req.session.userId = user.id; // log the user in before sending response
+    res.status(201).json({
+      message: "User created!",
+      user: {
+        name: user.name,
+        email: user.email,
+      }
+    })
+  } catch (error) {
+    if (error.name === "SequelizeValidationError") {
+      return res
+        .status(422)
+        .json({ errors: error.errors.map((e) => e.message) });
+    }
+    console.error(error);
+    res.status(500).json({
+      message: "Error occurred while creating a new user account"
+    })
+  }
+})
+
+app.post("/login", async (req, res) => {
+  try {
+    // find the user based on the email address in the body
+    const user = await User.findOne({ where: { email: req.body.email } });
+
+    if (user === null) {
+      return res.status(401).json({
+        message: "Incorrect credentials"
+      })
+    }
+
+    bcrypt.compare(req.body.password, user.password, (error, result) => {
+      if (result) {
+        // passwords match
+        req.session.userId = user.id;
+
+        res.status(200).json({
+          message: "Logged in successfully",
+          user: {
+            name: user.name,
+            email: user.email
+          }
+        })
+      } else {
+        // passwords don't match
+        return res.status(401).json({
+          message: "Incorrect credentials",
+        });
+      }
+    })
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ message: "An error occurred during the login process"})
+  }
+})
+
+app.delete("/logout", (req, res) => {
+  req.session.destroy((err) => {
+    if (err) {
+      return res.sendStatus(500);
+    }
+
+    res.clearCookie("connect.sid");
+    return res.sendStatus(200);
+  });
+});
+
 // Get all the jobs
-app.get("/jobs", async (req, res) => {
+app.get("/jobs", authenticateUser, async (req, res) => {
   try {
     const allJobs = await JobApplication.findAll();
 
@@ -31,7 +128,7 @@ app.get("/jobs", async (req, res) => {
 });
 
 // Get a specific job
-app.get("/jobs/:id", async (req, res) => {
+app.get("/jobs/:id", authenticateUser, async (req, res) => {
   const jobId = parseInt(req.params.id, 10);
 
   try {
@@ -49,7 +146,7 @@ app.get("/jobs/:id", async (req, res) => {
 });
 
 // Create a new job
-app.post("/jobs", async (req, res) => {
+app.post("/jobs", authenticateUser, async (req, res) => {
   try {
     const newJob = await JobApplication.create(req.body);
 
@@ -64,10 +161,16 @@ app.post("/jobs", async (req, res) => {
 });
 
 // Update a specific job
-app.patch("/jobs/:id", async (req, res) => {
+app.patch("/jobs/:id", authenticateUser, async (req, res) => {
   const jobId = parseInt(req.params.id, 10);
 
   try {
+    const record = await JobApplication.findOne({ where: { id: jobId } });
+    if (record && record.UserId !== parseInt(req.session.userId, 10)) {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to perform that action." });
+    }
     const [numberOfAffectedRows, affectedRows] = await JobApplication.update(
       req.body,
       { where: { id: jobId }, returning: true }
@@ -88,10 +191,16 @@ app.patch("/jobs/:id", async (req, res) => {
 });
 
 // Delete a specific job
-app.delete("/jobs/:id", async (req, res) => {
+app.delete("/jobs/:id", authenticateUser, async (req, res) => {
   const jobId = parseInt(req.params.id, 10);
 
   try {
+    const record = await JobApplication.findOne({ where: { id: jobId } });
+    if (record && record.UserId !== parseInt(req.session.userId, 10)) {
+      return res
+        .status(403)
+        .json({ message: "You are not authorized to perform that action." });
+    }
     const deleteOp = await JobApplication.destroy({ where: { id: jobId } });
 
     if (deleteOp > 0) {
